requirements.txt
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.1
Flask==2.0.3
future==0.18.2
idna==3.3
importlib-metadata==4.11.1
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
netmiko==3.4.0
ntc-templates==3.0.0
paramiko==2.9.2
pycparser==2.21
PyNaCl==1.5.0
pyserial==3.5
PyYAML==6.0
requests==2.27.1
scp==0.14.4
six==1.16.0
tenacity==8.0.1
textfsm==1.1.2
typing_extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
zipp==3.7.0